backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
asyncpg==0.29.0
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.2.1
alembic==1.13.1
psycopg2-binary==2.9.9
python-multipart==0.0.9
httpx==0.26.0